Add Place lookup screen, displaying all its relationships

+++ app/src/main/java/ly/david/mbjc/data/network/PlaceMusicBrainzModel.kt
@@ -24,8 +24,8 @@ internal data class PlaceMusicBrainzModel(
 
 internal data class Coordinates(
     @Json(name = "longitude")
-    val longitude: Float,
+    val longitude: Double,
 
     @Json(name = "latitude")
-    val latitude: Float,
+    val latitude: Double,
 )
